+++ bundle/Exception/ActionFailedException.php
@@ -6,5 +6,16 @@ use RuntimeException;
 
 class ActionFailedException extends RuntimeException
 {
+    /**
+     * ActionFailedException constructor.
+     *
+     * @param string $action
+     * @param string $reason
+     */
+    public function __construct($action, $reason)
+    {
+        $message = "InformationCollection action {$action} failed with reason {$reason}";
 
+        parent::__construct($message);
+    }
 }