+++ bundle/Exception/EmailNotSentException.php
@@ -6,5 +6,16 @@ use RuntimeException;
 
 class EmailNotSentException extends RuntimeException
 {
+    /**
+     * EmailNotSentException constructor.
+     *
+     * @param string $what
+     * @param string $why
+     */
+    public function __construct($what, $why)
+    {
+        $message = "Error occurred while trying to send email: {$what} failed with error {$why}";
 
+        parent::__construct($message);
+    }
 }